requirements.txt
flask
flask-socketio
gitpython
python-dotenv
docker
PyYAML
